+++ Runtime/RenderObject/Mesh.cpp
@@ -50,13 +50,13 @@ void Mesh::SetName(std::string_view name) {
     _pGpuMeshData->SetName(name);
 }
 
-void Mesh::SetIndices(ReadonlyArraySpan<uint16_t> indices) {
+void Mesh::SetIndices(ReadonlyArraySpan<uint32_t> indices) {
     Assert(indices.Count() == _pCpuMeshData->GetIndexCount());
     auto ptr = _pCpuMeshData->GetIndicesBegin();
     std::memcpy(ptr, indices.Data(), sizeof(uint16_t) * indices.Count());
 }
 
-void Mesh::SetIndices(ReadonlyArraySpan<int16_t> indices) {
+void Mesh::SetIndices(ReadonlyArraySpan<int32_t> indices) {
     Assert(indices.Count() == _pCpuMeshData->GetIndexCount());
     auto ptr = _pCpuMeshData->GetIndicesBegin();
     std::memcpy(ptr, indices.Data(), sizeof(uint16_t) * indices.Count());
